requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blis==0.7.10
catalogue==2.1.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confection==0.1.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
EasyProcess==1.1
elevenlabs==0.2.26
entrypoint2==1.1
et-xmlfile==1.1.0
executing==2.0.1
fonttools==4.44.3
frozenlist==1.4.0
idna==3.4
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
kiwisolver==1.4.5
langcodes==3.3.0
lml==0.1.0
load-dotenv==0.1.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
murmurhash==1.0.10
numpy==1.26.2
openai==1.3.3
openpyxl==3.1.2
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.3
parso==0.8.3
pathy==0.10.3
pickleshare==0.7.5
Pillow==10.1.0
preshed==4.0.0
prompt-toolkit==3.0.41
pure-eval==0.2.2
PyAutoGUI==0.9.54
pydantic==2.5.1
pydub==0.25.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.17.1
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
requests==2.31.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
texttable==1.7.0
thinc==8.2.1
tk==0.1.0
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wasabi==1.1.2
wcwidth==0.2.10
websockets==12.0
yarl==1.9.2
